jarvis/requirements.txt
pvporcupine==3.0.1
pyaudio==0.2.14
groq==0.15.0
huggingface-hub==0.27.1
pyttsx3==2.90
psycopg2-binary==2.9.9
duckduckgo-search==7.3.2
python-dotenv==1.0.1
SpeechRecognition==3.10.4
openwakeword==0.7.1
onnxruntime==1.17.1
tflite-runtime==2.14.0
pywhatkit==5.4
sqlalchemy==2.0.4
langdetect==1.0.9
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
firecrawl-py>=0.0.5
google-cloud-texttospeech>=2.14.1
yfinance>=0.2.36
